BackEnd/web/requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
sqlmodel
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]
asyncpg
bcrypt==4.0.1
redis
python-socketio
httpx